re-organise components neatly
